backend/requirements.txt
# CampusNexus Backend Dependencies

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
py-algorand-sdk>=2.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
authlib>=1.3.0
passlib>=1.7.4
python-multipart>=0.0.6
itsdangerous>=2.1.0
email-validator>=2.0.0

# AI/ML Dependencies for Semantic Skill Matching
sentence-transformers>=2.2.2
torch>=2.0.0

# IPFS/Pinata Integration
requests>=2.31.0
